requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-ipware==4.0.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
